# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "myExtBot API",
    "version": "0.1.0",
    "description": "REST API for myExtBot — an extensible multi-agent orchestration platform.\n\nAll endpoints under `/api/*` require an `x-api-key` header (set `MYEXTBOT_API_KEY` env var, default: `dev-key`).\n\nInteractive docs: `/api-docs` · Raw spec: `/api-docs/json`"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string"
          }
        }
      },
      "SuccessOk": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
      "PipelineStep": {
        "type": "object",
        "required": [
          "agentId",
          "toolName"
        ],
        "properties": {
          "agentId": {
            "type": "string"
          },
          "toolName": {
            "type": "string"
          },
          "arguments": {
            "type": "object"
          }
        }
      },
      "AgentPipeline": {
        "type": "object",
        "required": [
          "id",
          "name",
          "steps"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "steps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PipelineStep"
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
...
  "paths": {
...
    "/api/pipelines": {
      "get": {
        "tags": [
          "Pipelines"
        ],
        "summary": "List all registered pipelines",
        "responses": {
          "200": {
            "description": "Array of AgentPipeline objects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AgentPipeline"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Pipelines"
        ],
        "summary": "Register a new agent pipeline",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentPipeline"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pipeline registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "pipeline": {
                      "$ref": "#/components/schemas/AgentPipeline"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/pipelines/{id}": {
      "get": {
        "tags": [
          "Pipelines"
        ],
        "summary": "Get a pipeline by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The pipeline object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentPipeline"
                }
              }
            }
          },
          "404": {
            "description": "Pipeline not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Pipelines"
        ],
        "summary": "Unregister (delete) a pipeline",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pipeline deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessOk"
                }
              }
            }
          },
          "404": {
            "description": "Pipeline not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/pipelines/{id}/run": {
      "post": {
        "tags": [
          "Pipelines"
        ],
        "summary": "Execute a pipeline",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "initialInput": {
                    "type": "object",
                    "description": "Initial input data passed to the first step"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pipeline run result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "stepResults": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Pipeline not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/agents/route": {
      "get": {
        "tags": [
          "Agents"
        ],
        "summary": "Route a query to the most relevant agents",
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Natural language query to route"
          },
          {
            "in": "query",
            "name": "topN",
            "schema": {
              "type": "integer",
              "default": 3
            },
            "description": "Number of top suggestions to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Array of agent routing suggestions (sorted by score)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "agentId": {
                        "type": "string"
                      },
                      "score": {
                        "type": "number"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing query parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}